+++ __tests__/pages/resource/id/index.test.tsx
